httpserver/src/test/java/httpserver/SimpleHTTPServerTest.java
package httpserver;

import org.hamcrest.core.Is;
import org.junit.Test;

import java.io.ByteArrayOutputStream;
import java.io.IOException;
import java.io.InputStream;

import static org.junit.Assert.assertThat;

/**
 * Tests for methods of SimpleHTTPServer class
 */
public class SimpleHTTPServerTest {
    private SimpleHTTPServer server = new SimpleHTTPServer();
    private String[] args = {"8080"};

////  For running server in local environment for tests
////  todo: comment this test before pushing
//    @Test
//    public void test0() throws Exception {
//        String[] args = {"8080"};
//        int expected = server.openPort(args);
//    }

    @Test
    public void strArgsFail() throws Exception {
        String[] args = {"asd"};
        int expected = server.openPort(args);
        assertThat(expected, Is.is(-1));
    }

    @Test
    public void portWrongRangeFail() throws Exception {
        String[] args = {"80"};
        int expected = server.openPort(args);
        assertThat(expected, Is.is(0));
    }

    @Test
    public void noArgsFail() throws Exception {
        String[] args = {};
        int expected = server.openPort(args);
        assertThat(expected, Is.is(0));
        args = null;
        expected = server.openPort(args);
        assertThat(expected, Is.is(0));
    }

    private byte[] readInputStream(InputStream is) {
        try (ByteArrayOutputStream os = new ByteArrayOutputStream()) {
            byte[] buffer = new byte[0xFFFF];
            for (int len; (len = is.read(buffer)) != -1; )
                os.write(buffer, 0, len);
            os.flush();
            return os.toByteArray();
        } catch (IOException e) {
            e.printStackTrace();
            return null;
        }
    }

}